Add data split feature to compare compression on subparts
Split dataset on z-axis to break data into smaller chunks
and compress them individually to compare compression on
the chunks.

Update mpi executor to iterate across the split arrays
with the config iterator adding an index value as one
of the args.

Remove the visualization output since we evetually
want an non-uniform distribution.

Accept input arg to determine the number of splits
representing the processes across which data is split
leading to non-uniform distribution.